+++ android/src/main/java/io/sentry/react/RNSentryPackage.java
@@ -55,5 +55,4 @@ public class RNSentryPackage extends TurboReactPackage {
           new RNSentryOnDrawReporterManager(reactContext)
     );
   }
-
 }