+++ src/js/wrapper.ts
@@ -116,6 +116,7 @@ interface SentryNativeWrapper {
   getCurrentReplayId(): string | null;
 
   crashedLastRun(): Promise<boolean | null>;
+  getNewScreenTimeToDisplay(): Promise<number | null | undefined>;
 }
 
 const EOL = utf8ToBytes('\n');
@@ -687,6 +688,14 @@ export const NATIVE: SentryNativeWrapper = {
 
     const result = RNSentry.crashedLastRun();
     return typeof result === 'boolean' ? result : null;
+  },
+
+  getNewScreenTimeToDisplay(): Promise<number | null | undefined> {
+    if (!this.enableNative || !this._isModuleLoaded(RNSentry)) {
+      return Promise.resolve(null);
+    }
+
+    return RNSentry.getNewScreenTimeToDisplay();
   },
 
   /**